+++ plugins/module_utils/base_module.py
@@ -2,6 +2,8 @@
 # Simplified BSD License (see licenses/simplified_bsd.txt or https://opensource.org/licenses/BSD-2-Clause )
 from __future__ import absolute_import, division, print_function
 
+from copy import deepcopy
+
 from ansible.module_utils.basic import AnsibleModule, env_fallback
 
 __metaclass__ = type
@@ -12,17 +14,13 @@ class BaseModule(AnsibleModule):
 
     argument_spec = dict(
         nutanix_host=dict(
-            type="str",
-            fallback=(env_fallback, ["NUTANIX_HOST"]),
-            required=True,
+            type="str", fallback=(env_fallback, ["NUTANIX_HOST"]), required=True
         ),
         nutanix_port=dict(
             default="9440", type="str", fallback=(env_fallback, ["NUTANIX_PORT"])
         ),
         nutanix_username=dict(
-            type="str",
-            fallback=(env_fallback, ["NUTANIX_USERNAME"]),
-            required=True,
+            type="str", fallback=(env_fallback, ["NUTANIX_USERNAME"]), required=True
         ),
         nutanix_password=dict(
             type="str",
@@ -38,10 +36,10 @@ class BaseModule(AnsibleModule):
     )
 
     def __init__(self, **kwargs):
+        argument_spec = deepcopy(self.argument_spec)
         if kwargs.get("argument_spec"):
-            kwargs["argument_spec"].update(self.argument_spec)
-        else:
-            kwargs["argument_spec"] = self.argument_spec
+            argument_spec.update(kwargs["argument_spec"])
+        kwargs["argument_spec"] = argument_spec
 
         if not kwargs.get("supports_check_mode"):
             kwargs["supports_check_mode"] = True