+++ plugins/module_utils/prism/groups.py
@@ -12,12 +12,31 @@ class Groups(Prism):
         resource_type = "/groups"
         super(Groups, self).__init__(module, resource_type=resource_type)
 
-    def get_uuid(self, value, key="name", entity_type=""):
+    def get_uuid(
+        self, value, key="name", entity_type="", raise_error=True, no_response=False
+    ):
         data = {
             "entity_type": entity_type,
             "filter_criteria": "{0}=={1}".format(key, value),
         }
-        resp = self.list(data, use_base_url=True)
+        resp = self.list(
+            data, use_base_url=True, raise_error=raise_error, no_response=no_response
+        )
         if resp.get("group_results"):
             return resp["group_results"][0]["entity_results"][0]["entity_id"]
         return None
+
+
+def get_entity_uuid(config, module, key, entity_type):
+    if "name" in config:
+        groups = Groups(module)
+        name = config["name"]
+        uuid = groups.get_uuid(value=name, key=key, entity_type=entity_type)
+        if not uuid:
+            entity_type = entity_type.replace("_", " ")
+            error = "{0} {1} not found.".format(entity_type.capitalize(), name)
+            return None, error
+    elif "uuid" in config:
+        uuid = config["uuid"]
+
+    return uuid, None