+++ plugins/modules/ntnx_vms.py
@@ -10,109 +10,58 @@ __metaclass__ = type
 DOCUMENTATION = r"""
 ---
 module: ntnx_vms
-short_description: VM module which suports VM CRUD operations
+short_description: VM module which supports VM CRUD states
 version_added: 1.0.0
-description: 'Create, Update, Delete, Power-on, Power-off Nutanix VM''s'
+description: "Create, Update, Delete, Power-on, Power-off Nutanix VM's"
 options:
-  name:
-    description: VM Name
+  state:
+    description:
+      - Specify state
+      - If C(state) is set to C(present) then the opperation will be  create the item
+      - >-
+        If C(state) is set to C(absent) and if the item exists, then
+        item is removed.
+      - If C(state) is set to C(power_on) then the opperation will be  power on the VM
+      - If C(state) is set to C(power_off) then the opperation will be  power off the VM
+      - If C(state) is set to C(soft_shutdown) then the opperation will be  soft shutdown  the VM
+      - If C(state) is set to C(hard_poweroff) then the opperation will be  hard poweroff  the VM
+    choices:
+      - present
+      - absent
+      - power_on
+      - power_off
+      - soft_shutdown
+      - hard_poweroff
+    type: str
+    default: present
+  wait:
+    description: Wait for the  CRUD operation to complete.
+    type: bool
+    required: false
+    default: True
+  desc:
+    description: A description for VM.
     required: false
     type: str
   vm_uuid:
     description: VM UUID
     type: str
-  desc:
-    description: A description for VM.
-    required: false
-    type: str
-  project:
-    description: Name or UUID of the project
-    required: false
-    type: dict
-    suboptions:
-      name:
-        description:
-          - Project Name
-          - Mutually exclusive with C(uuid)
-        type: str
-      uuid:
-        description:
-          - Project UUID
-          - Mutually exclusive with C(name)
-        type: str
-  cluster:
-    description:
-      - Name or UUID of the cluster on which the VM will be placed
-    type: dict
-    required: false
-    suboptions:
-      name:
-        description:
-          - Cluster Name
-          - Mutually exclusive with C(uuid)
-        type: str
-      uuid:
-        description:
-          - Cluster UUID
-          - Mutually exclusive with C(name)
-        type: str
-  vcpus:
-    description:
-      - Number of sockets
-    required: false
-    type: int
-    default: 1
-  cores_per_vcpu:
-    description:
-      - This is the number of vcpus per socket
-    required: false
-    type: int
-    default: 1
-  memory_gb:
-    description:
-      - Memory size in GB
-    required: false
-    type: int
-    default: 1
-  networks:
-    description:
-      - list of subnets to which the VM needs to connect to
-    type: list
-    elements: dict
-    required: false
-    suboptions:
-      subnet:
-        description:
-          - Name or UUID of the subnet to which the VM should be connnected
-        type: dict
-        suboptions:
-          name:
-            description:
-              - Subnet Name
-              - Mutually exclusive with C(uuid)
-            type: str
-          uuid:
-            description:
-              - Subnet UUID
-              - Mutually exclusive with C(name)
-            type: str
-      private_ip:
-        description:
-          - Optionally assign static IP to the VM
-        type: str
-        required: false
-      is_connected:
-        description:
-          - Connect or disconnect the VM to the subnet
-        type: bool
-        required: false
-        default: true
   disks:
     description:
       - List of disks attached to the VM
     type: list
     elements: dict
     suboptions:
+      uuid:
+        description:
+          - Disk's uuid
+        type: str
+      state:
+        description:
+          - Disk's state to delete it
+        type: str
+        choices:
+          - absent
       type:
         description:
           - CDROM or DISK
@@ -169,76 +118,14 @@ options:
         type: bool
         description:
           - Mutually exclusive with C(clone_image) and C(storage_container)
-  boot_config:
-    description:
-      - >-
-        Indicates whether the VM should use Secure boot, UEFI boot or Legacy
-        boot.
-    type: dict
-    required: false
-    suboptions:
-      boot_type:
-        description:
-          - Boot type of VM.
-        choices:
-          - LEGACY
-          - UEFI
-          - SECURE_BOOT
-        default: LEGACY
-        type: str
-      boot_order:
-        description:
-          - Applicable only for LEGACY boot_type
-          - Boot device order list
-        type: list
-        elements: str
-        default:
-          - CDROM
-          - DISK
-          - NETWORK
-  guest_customization:
-    description:
-      - cloud_init or sysprep guest customization
-    type: dict
-    suboptions:
-      type:
-        description:
-          - cloud_init or sysprep type
-        type: str
-        required: True
-        choices:
-          - cloud_init
-          - sysprep
-      script_path:
-        description:
-          - Absolute file path to the script.
-        type: path
-        required: true
-      is_overridable:
-        description:
-          - Flag to allow override of customization during deployment.
-        type: bool
-        default: false
-        required: false
-  timezone:
-    description:
-      - VM's hardware clock timezone in IANA TZDB format (America/Los_Angeles).
-    type: str
-    default: UTC
-    required: false
-  categories:
-    description:
-      - categories to be attached to the VM.
-    type: dict
-    required: false
 extends_documentation_fragment:
       - nutanix.ncp.ntnx_credentials
+      - nutanix.ncp.ntnx_vms_base
       - nutanix.ncp.ntnx_operations
 author:
  - Prem Karat (@premkarat)
  - Gevorg Khachatryan (@Gevorg-Khachatryan-97)
  - Alaa Bishtawi (@alaa-bish)
- - Dina AbuHijleh (@dina-abuhijleh)
 """
 
 EXAMPLES = r"""
@@ -393,6 +280,171 @@ EXAMPLES = r"""
       nutanix_password: "{{ password }}"
       validate_certs: False
       vm_uuid: '{{ vm_uuid }}'
+
+  - name: update vm by  values for memory, vcpus and cores_per_vcpu, timezone
+    ntnx_vms:
+      vm_uuid: "{{ vm.vm_uuid }}"
+      vcpus: 1
+      cores_per_vcpu: 1
+      memory_gb: 1
+      timezone: UTC
+
+  - name: Update VM by adding all type of  disks
+    ntnx_vms:
+      vm_uuid: "{{ vm.vm_uuid }}"
+      disks:
+        - type: "DISK"
+          clone_image:
+            name: "{{ ubuntu }}"
+          bus: "SCSI"
+          size_gb: 20
+        - type: DISK
+          size_gb: 1
+          bus: PCI
+        - type: "DISK"
+          size_gb: 1
+          bus: "SATA"
+        - type: "DISK"
+          size_gb: 1
+          bus: "SCSI"
+        - type: DISK
+          size_gb: 1
+          bus: SCSI
+          storage_container:
+            uuid: "{{ storage_container.uuid }}"
+        - type: "DISK"
+          bus: "IDE"
+          size_gb: 1
+        - type: "CDROM"
+          bus: "IDE"
+          empty_cdrom: True
+
+  - name: Update VM by increasing  the size of the disks
+    ntnx_vms:
+      vm_uuid: "{{ result.vm_uuid }}"
+      disks:
+        - type: "DISK"
+          uuid: "{{ result.response.spec.resources.disk_list[0].uuid }}"
+          size_gb: 22
+        - type: DISK
+          uuid: "{{ result.response.spec.resources.disk_list[1].uuid }}"
+          size_gb: 2
+        - type: "DISK"
+          uuid: "{{ result.response.spec.resources.disk_list[2].uuid }}"
+          size_gb: 2
+        - type: "DISK"
+          size_gb: 2
+          uuid: "{{ result.response.spec.resources.disk_list[3].uuid }}"
+        - type: DISK
+          size_gb: 2
+          uuid: "{{ result.response.spec.resources.disk_list[4].uuid }}"
+        - type: "DISK"
+          uuid: "{{ result.response.spec.resources.disk_list[5].uuid }}"
+          size_gb: 1
+
+  - name: Update VM by removing all type of  disks
+    ntnx_vms:
+      vm_uuid: "{{ result.vm_uuid }}"
+      disks:
+        - state: absent
+          uuid: "{{ result.response.spec.resources.disk_list[0].uuid }}"
+        - state: absent
+          uuid: "{{ result.response.spec.resources.disk_list[1].uuid }}"
+        - state: absent
+          uuid: "{{ result.response.spec.resources.disk_list[2].uuid }}"
+        - state: absent
+          uuid: "{{ result.response.spec.resources.disk_list[3].uuid }}"
+        - state: absent
+          uuid: "{{ result.response.spec.resources.disk_list[4].uuid }}"
+        - state: absent
+          uuid: "{{ result.response.spec.resources.disk_list[5].uuid }}"
+        - state: absent
+          uuid: "{{ result.response.spec.resources.disk_list[6].uuid }}"
+
+  - name: Update VM by adding subnets
+    ntnx_vms:
+      vm_uuid: "{{ result.vm_uuid }}"
+      networks:
+        - is_connected: true
+          subnet:
+            uuid: "{{ network.dhcp.uuid }}"
+        - is_connected: false
+          subnet:
+            uuid: "{{ static.uuid }}"
+          private_ip: "{{ network.static.ip }}"
+
+  - name: Update VM by editing a subnet is_connected
+    ntnx_vms:
+      vm_uuid: "{{ vm.vm_uuid }}"
+      desc: disconnect and connects nic's
+      networks:
+        - is_connected: true
+          uuid: "{{ result.response.spec.resources.nic_list[1].uuid }}"
+        - is_connected: false
+          uuid: "{{ result.response.spec.resources.nic_list[0].uuid }}"
+
+  - name: Update VM by change the private ip for subnet
+    ntnx_vms:
+      vm_uuid: "{{ vm.vm_uuid }}"
+      name: updated
+      desc: change ip
+      networks:
+        - is_connected: true
+          private_ip: "10.30.30.79"
+          uuid: "{{ result.response.spec.resources.nic_list[1].uuid }}"
+
+  - name: Update VM by change vlan subnet
+    ntnx_vms:
+      vm_uuid: "{{ vm.vm_uuid }}"
+      name: updated
+      desc: change vlan
+      categories:
+        AppType:
+          - Apache_Spark
+      networks:
+        - is_connected: false
+          subnet:
+            name: vlan1211
+          uuid: "{{ result.response.spec.resources.nic_list[0].uuid }}"
+
+  - name: Update VM by deleting a subnet
+    ntnx_vms:
+      vm_uuid: "{{ vm.vm_uuid }}"
+      networks:
+        - state: absent
+          uuid: "{{ result.response.spec.resources.nic_list[0].uuid }}"
+        - state: absent
+          uuid: "{{ result.response.spec.resources.nic_list[1].uuid }}"
+
+  - name: hard power off the vm
+    ntnx_vms:
+        vm_uuid: "{{ vm.vm_uuid }}"
+        state: hard_poweroff
+    register: result
+    ignore_errors: true
+  - name: power on the vm
+    ntnx_vms:
+        state: power_on
+        vm_uuid: "{{ vm.vm_uuid }}"
+
+  - name: soft shut down the vm
+    ntnx_vms:
+        state: soft_shutdown
+        vm_uuid: "{{ vm.vm_uuid }}"
+
+  - name: Create VM with minimum requiremnts with hard_poweroff opperation
+    ntnx_vms:
+        state: hard_poweroff
+        name: integration_test_opperations_vm
+        cluster:
+          name: "{{ cluster.name }}"
+
+  - name: Create VM with minimum requiremnts with poweroff opperation
+    ntnx_vms:
+        state: power_off
+        name: integration_test_opperations_vm
+        cluster:
+          name: "{{ cluster.name }}"
 """
 
 RETURN = r"""
@@ -680,27 +732,27 @@ task_uuid:
   sample: "82c5c1d3-eb6a-406a-8f58-306028099d21"
 """
 
+
+from copy import deepcopy  # noqa: E402
+
+from ..module_utils import utils  # noqa: E402
 from ..module_utils.base_module import BaseModule  # noqa: E402
+from ..module_utils.prism.spec.vms import DefaultVMSpec  # noqa: E402
 from ..module_utils.prism.tasks import Task  # noqa: E402
 from ..module_utils.prism.vms import VM  # noqa: E402
-from ..module_utils.utils import remove_param_with_none_value  # noqa: E402
 
 
 def get_module_spec():
+    default_vm_spec = deepcopy(DefaultVMSpec.vm_argument_spec)
+
     mutually_exclusive = [("name", "uuid")]
 
     entity_by_spec = dict(name=dict(type="str"), uuid=dict(type="str"))
 
-    network_spec = dict(
-        subnet=dict(
-            type="dict", options=entity_by_spec, mutually_exclusive=mutually_exclusive
-        ),
-        private_ip=dict(type="str", required=False),
-        is_connected=dict(type="bool", default=True),
-    )
-
     disk_spec = dict(
         type=dict(type="str", choices=["CDROM", "DISK"], default="DISK"),
+        uuid=dict(type="str"),
+        state=dict(type="str", choices=["absent"]),
         size_gb=dict(type="int"),
         bus=dict(type="str", choices=["SCSI", "PCI", "SATA", "IDE"], default="SCSI"),
         storage_container=dict(
@@ -712,35 +764,20 @@ def get_module_spec():
         empty_cdrom=dict(type="bool"),
     )
 
-    boot_config_spec = dict(
-        boot_type=dict(
-            type="str", choices=["LEGACY", "UEFI", "SECURE_BOOT"], default="LEGACY"
-        ),
-        boot_order=dict(
-            type="list", elements="str", default=["CDROM", "DISK", "NETWORK"]
-        ),
-    )
-
-    gc_spec = dict(
-        type=dict(type="str", choices=["cloud_init", "sysprep"], required=True),
-        script_path=dict(type="path", required=True),
-        is_overridable=dict(type="bool", default=False),
-    )
-
     module_args = dict(
-        name=dict(type="str", required=False),
-        vm_uuid=dict(type="str"),
+        state=dict(
+            type="str",
+            choices=[
+                "present",
+                "absent",
+                "power_on",
+                "power_off",
+                "soft_shutdown",
+                "hard_poweroff",
+            ],
+            default="present",
+        ),
         desc=dict(type="str"),
-        project=dict(
-            type="dict", options=entity_by_spec, mutually_exclusive=mutually_exclusive
-        ),
-        cluster=dict(
-            type="dict", options=entity_by_spec, mutually_exclusive=mutually_exclusive
-        ),
-        vcpus=dict(type="int", default=1),
-        cores_per_vcpu=dict(type="int", default=1),
-        memory_gb=dict(type="int", default=1),
-        networks=dict(type="list", elements="dict", options=network_spec),
         disks=dict(
             type="list",
             elements="dict",
@@ -748,19 +785,17 @@ def get_module_spec():
             mutually_exclusive=[
                 ("storage_container", "clone_image", "empty_cdrom"),
                 ("size_gb", "empty_cdrom"),
+                ("uuid", "bus"),
             ],
         ),
-        boot_config=dict(type="dict", options=boot_config_spec),
-        guest_customization=dict(type="dict", options=gc_spec),
-        timezone=dict(type="str", default="UTC"),
-        categories=dict(type="dict"),
     )
-
-    return module_args
+    default_vm_spec.update(module_args)
+    return default_vm_spec
 
 
 def create_vm(module, result):
     vm = VM(module)
+    state = module.params.get("state")
     spec, error = vm.get_spec()
     if error:
         result["error"] = error
@@ -777,10 +812,124 @@ def create_vm(module, result):
     result["vm_uuid"] = vm_uuid
     result["task_uuid"] = resp["status"]["execution_context"]["task_uuid"]
 
-    if module.params.get("wait"):
+    if module.params.get("wait") or state in [
+        "soft_shutdown",
+        "hard_poweroff",
+        "power_off",
+    ]:
         wait_for_task_completion(module, result)
         resp = vm.read(vm_uuid)
+        spec = resp.copy()
+        spec.pop("status")
         result["response"] = resp
+
+    if state == "soft_shutdown":
+        resp = vm.soft_shutdown(spec)
+    elif state in ["hard_poweroff", "power_off"]:
+        resp = vm.hard_power_off(spec)
+
+    if state in ["soft_shutdown", "hard_poweroff", "power_off"]:
+        result["response"] = resp
+        result["task_uuid"] = resp["status"]["execution_context"]["task_uuid"]
+        if module.params.get("wait"):
+            wait_for_task_completion(module, result, False)
+            state = result["response"].get("status")
+            resp = vm.read(vm_uuid)
+            result["response"] = resp
+            if state == "FAILED":
+                error = "VM 'soft_shutdown' state failed, use 'hard_poweroff' instead"
+                result["error"] = error
+                module.fail_json(msg=error, **result)
+
+
+def update_vm(module, result):
+    vm_uuid = module.params["vm_uuid"]
+    state = module.params.get("state")
+
+    vm = VM(module)
+    resp = vm.read(vm_uuid)
+    result["response"] = resp
+    utils.strip_extra_attrs_from_status(resp["status"], resp["spec"])
+    resp["spec"] = resp.pop("status")
+    spec, error = vm.get_spec(resp)
+
+    if error:
+        result["error"] = error
+        module.fail_json(msg="Failed generating VM Spec", **result)
+
+    if module.check_mode:
+        result["response"] = spec
+        return
+
+    if utils.check_for_idempotency(spec, resp, state=state):
+        result["skipped"] = True
+        module.exit_json(msg="Nothing to change")
+
+    is_vm_on = vm.is_on(spec)
+    is_powered_off = False
+
+    if is_vm_on and vm.is_restart_required():
+        if not module.params.get("force_power_off"):
+            module.fail_json(
+                "To make these changes, the VM should be restarted, but 'force_power_off' is False"
+            )
+
+        power_off_vm(vm, module, result)
+        vm.update_entity_spec_version(spec)
+        vm.set_power_state(spec, "OFF")
+        is_powered_off = True
+
+    resp = vm.update(spec, vm_uuid)
+    spec = resp.copy()
+    spec.pop("status")
+
+    result["changed"] = True
+    result["response"] = resp
+    result["vm_uuid"] = vm_uuid
+    result["task_uuid"] = resp["status"]["execution_context"]["task_uuid"]
+
+    if (
+        module.params.get("wait")
+        or is_powered_off
+        or (is_vm_on and state in ["soft_shutdown", "hard_poweroff", "power_off"])
+    ):
+        wait_for_task_completion(module, result)
+        resp = vm.read(vm_uuid)
+        spec = resp.copy()
+        spec.pop("status")
+        result["response"] = resp
+
+    wait = False
+    if state == "soft_shutdown" and is_vm_on and not is_powered_off:
+        resp = vm.soft_shutdown(spec)
+        wait = True
+    elif state in ["hard_poweroff", "power_off"] and is_vm_on and not is_powered_off:
+        resp = vm.hard_power_off(spec)
+        wait = True
+    elif is_powered_off or (state == "power_on" and not is_vm_on):
+        resp = vm.power_on(spec)
+        wait = True
+
+    if wait:
+        result["response"] = resp
+        result["task_uuid"] = resp["status"]["execution_context"]["task_uuid"]
+        if module.params.get("wait"):
+            wait_for_task_completion(module, result, False)
+            response_state = result["response"].get("status")
+            if response_state == "FAILED":
+                result[
+                    "warning"
+                ] = "VM 'soft_shutdown' operation failed, use 'hard_poweroff' instead"
+
+            resp = vm.read(vm_uuid)
+            result["response"] = resp
+
+
+def power_off_vm(vm, module, result):
+    resp = vm.hard_power_off(result["response"])
+    result["task_uuid"] = resp["status"]["execution_context"]["task_uuid"]
+    wait_for_task_completion(module, result)
+    result["response"] = resp
 
 
 def delete_vm(module, result):
@@ -800,23 +949,22 @@ def delete_vm(module, result):
         wait_for_task_completion(module, result)
 
 
-def wait_for_task_completion(module, result):
+def wait_for_task_completion(module, result, raise_error=True):
     task = Task(module)
     task_uuid = result["task_uuid"]
-    resp = task.wait_for_completion(task_uuid)
+    resp = task.wait_for_completion(task_uuid, raise_error=raise_error)
     result["response"] = resp
+    if not result.get("vm_uuid") and resp.get("entity_reference_list"):
+        result["vm_uuid"] = resp["entity_reference_list"][0]["uuid"]
 
 
 def run_module():
     module = BaseModule(
         argument_spec=get_module_spec(),
         supports_check_mode=True,
-        required_if=[
-            ("state", "present", ("name",)),
-            ("state", "absent", ("vm_uuid",)),
-        ],
+        required_if=[("vm_uuid", None, ("name",)), ("state", "absent", ("vm_uuid",))],
     )
-    remove_param_with_none_value(module.params)
+    utils.remove_param_with_none_value(module.params)
     result = {
         "changed": False,
         "error": None,
@@ -825,10 +973,13 @@ def run_module():
         "task_uuid": None,
     }
     state = module.params["state"]
-    if state == "present":
-        create_vm(module, result)
-    elif state == "absent":
+
+    if state == "absent":
         delete_vm(module, result)
+    elif module.params.get("vm_uuid"):
+        update_vm(module, result)
+    else:
+        create_vm(module, result)
 
     module.exit_json(**result)